rename payload to record
